requirements.txt
sqlglot>=25.0.0
PyYAML>=6.0.0
